+++ cardano-base/CHANGELOG.md
@@ -1,8 +1,9 @@
 # Changelog for `cardano-base`
 
-## 0.1.1.1
+## 0.1.2.0
 
-*
+* Add `Cardano.Base.IP` module with `IPv4` and `IPv6` newtype wrappers to avoid orphan instances and laziness
+* Add `testlib` for `Arbitrary` and `ToExpr` instances
 
 ## 0.1.1.0